+++ NEWS.md
@@ -14,6 +14,8 @@
   of arguments currently tested is shown as much as possible (#11).
 - Fix a crash in the strange case when `conditionMessage()` produces more than
   one error message (#12).
+- If multiple warnings are generated, they are all returned in a concatenated
+  string, instead of returning only the last one (#14).
 
 # CBTF 0.6.0 (2026-02-13)